+++ src/uiobjects/uiimglist.cpp
@@ -4,10 +4,10 @@
  * and open the template in the editor.
  */
 
-/* 
+/*
  * File:   UIImgList.cpp
  * Author: Ryuk
- * 
+ *
  * Created on 4 de abril de 2020, 11:46
  */
 
@@ -24,23 +24,23 @@ UIImgList::UIImgList() {
 }
 
 UIImgList::~UIImgList() {
-    
+
 }
 
 /**
  * Overriding the parent method
- * @return 
+ * @return
  */
 unsigned int UIImgList::getElemVisibles(){
     unsigned int ret = 0;
-    
+
     if (mode == IMGTHUMBMODE && prevImgWidth > 0 && prevImgHeight > 0){
 //        int horizontalElems = this->getW() / prevImgWidth;
 //        int verticalElems = this->getH() / prevImgHeight;
-        
+
         int horizontalElems = getImgGestor()->calcMaxX(prevImgWidth, this->getW());
         int verticalElems = getImgGestor()->calcMaxY(prevImgHeight, this->getH());
-        
+
         ret = horizontalElems * verticalElems;
     } else if (mode == IMGGROUPMODE){
         ret = UIListGroup::getElemVisibles();
@@ -49,7 +49,7 @@ unsigned int UIImgList::getElemVisibles(){
 }
 
 /**
- * 
+ *
  * @param evento
  */
 void UIImgList::checkPos(tEvento evento){
@@ -61,7 +61,7 @@ void UIImgList::checkPos(tEvento evento){
         t_region regionPantalla = {this->getX(), this->getY(), this->getW(), this->getH()};
         int posPulsada = getImgGestor()->getPosThumb(mx - this->getX(), my - this->getY(), prevImgWidth, prevImgHeight, regionPantalla);
         this->setPosActualLista(posPulsada + this->getPosIniLista());
-        
+
         //Se setea el checked cuando se hace doble click
         if (evento.isMousedblClick){
             if (lastSelectedPos == this->posActualLista){
@@ -77,18 +77,18 @@ void UIImgList::checkPos(tEvento evento){
 }
 
 /**
- * 
+ *
  * @param evento
  */
 void UIImgList::action(tEvento *evento){
     if (popup == false && mode == IMGTHUMBMODE && prevImgWidth > 0 && prevImgHeight > 0){
         if (evento->resize || evento->isKey || evento->isJoy || (evento->isMouse &&
-                                               (evento->mouse == MOUSE_BUTTON_WHEELDOWN || evento->mouse == MOUSE_BUTTON_WHEELUP) 
+                                               (evento->mouse == MOUSE_BUTTON_WHEELDOWN || evento->mouse == MOUSE_BUTTON_WHEELUP)
                                                 && evento->mouse_state == 1
                 ) ){
-                                         
+
             Traza::print("UIImgList::action: " + this->getName(), W_DEBUG);
-            
+
             if ((evento->key == SDLK_LEFT || evento->joy == JoyMapper::getJoyMapper(JOY_BUTTON_LEFT))){
                this->setImgDrawed(prevElement(1));
                reloadImages = !this->getImgDrawed();
@@ -120,7 +120,7 @@ void UIImgList::action(tEvento *evento){
             } else {
                 UIListCommon::action(evento);
             }
-            
+
         } else if (evento->isMouse && evento->mouse == MOUSE_BUTTON_LEFT && evento->mouse_state == SDL_PRESSED){
             Traza::print("UIImgList::action: Mouse Pressed: " + this->getName(), W_DEBUG);
             checkPos(*evento);
@@ -132,73 +132,73 @@ void UIImgList::action(tEvento *evento){
         }
     } else if (mode == IMGGROUPMODE){
         UIListGroup::action(evento);
-    } 
+    }
 }
 
 /**
- * 
+ *
  * @param evento
  */
 void UIImgList::eventoResize(tEvento *evento){
     int horizontalElems = getImgGestor()->calcMaxX(prevImgWidth, this->getW());
     int verticalElems = getImgGestor()->calcMaxY(prevImgHeight, this->getH());
     int total = horizontalElems * verticalElems;
-    
-    Traza::print(string("Redimensionando a w: ") + Constant::TipoToStr(this->getW()) 
+
+    Traza::print(string("Redimensionando a w: ") + Constant::TipoToStr(this->getW())
         + string(", h: ") + Constant::TipoToStr(this->getH()), W_DEBUG);
-    
+
     int ini = 0;
     int fin = 0;
-    
+
     if (horizontalElems > 0){
-        
+
         ini = getPosActualLista();
         fin = getPosActualLista();
-        
+
         do{
             if (ini > 0){
                 ini--;
             }
-            if (fin - ini < total - 1 && fin < this->getSize() - 1){
+            if (fin - ini < total - 1 && fin < (int)this->getSize() - 1){
                 fin++;
             }
-        } while (fin - ini < total - 1 && fin - ini < this->getSize() - 1);
-          
+        } while (fin - ini < total - 1 && fin - ini < (int)this->getSize() - 1);
+
         int resto = 0;
-        
+
 //        if (fin >= (this->getSize() / horizontalElems) * horizontalElems &&
 //            fin < (this->getSize() / horizontalElems) * horizontalElems + horizontalElems - 1
 //            && ini > 0){
 //            resto = fin % horizontalElems;
 //        }
-        
-        while(ini % horizontalElems != 0 && ini < getPosActualLista()){
+
+        while(ini % horizontalElems != 0 && ini < (int)getPosActualLista()){
             ini++;
-            if (fin < getSize() - 1) 
+            if (fin < (int)getSize() - 1)
                 fin++;
         }
-        
+
 
         setPosIniLista(ini);
         setPosFinLista(fin);
-        
+
         Traza::print("resto: ", resto, W_DEBUG);
-    
+
 //    double percentTotal = getPosFinLista() - getPosIniLista();
 //    double percentLeft = 0;
 //    double percentRight = 0;
-//    
+//
 //    if (percentTotal > 1){
 //        percentLeft = (getPosActualLista() - getPosIniLista()) / percentTotal;
 //        percentRight = (getPosFinLista() - getPosActualLista()) / percentTotal;
 //        int elemsLeft = percentLeft * total - 1;
 //        int elemsRight = percentRight * total;
-//        
+//
 //        int ini = getPosActualLista() > elemsLeft ? getPosActualLista() - elemsLeft : 0;
 //        int fin = ini + total - 1 < this->getSize() ? ini + total - 1 : this->getSize() - 1;
 //        int resto = fin % horizontalElems;
-//        
-//        
+//
+//
 ////        if (fin == this->getSize() - 1 && ini > 0 && getPosActualLista() > this->getSize() - horizontalElems + resto -1){
 ////            setPosIniLista(ini + resto);
 ////            setPosFinLista(fin);
@@ -206,23 +206,23 @@ void UIImgList::eventoResize(tEvento *evento){
 //            setPosIniLista(ini);
 //            setPosFinLista(fin);
 ////        }
-        
-        
+
+
     } else {
         setPosActualLista(0);
         setPosIniLista(0);
-        setPosFinLista(total > this->getSize() ? this->getSize() - 1 : total -1);
+        setPosFinLista(total > (int)this->getSize() ? this->getSize() - 1 : total -1);
     }
 }
 
 /**
- * 
- * @return 
+ *
+ * @return
  */
 bool UIImgList::prevPage(int positions){
     int verticalElems = positions > 0 ? positions : getImgGestor()->calcMaxY(prevImgHeight, this->getH());
     unsigned int posIni = getPosIniLista();
-    
+
     if (verticalElems > 0){
         for (int i=0; i < verticalElems; i++){
             prevRow();
@@ -232,13 +232,13 @@ bool UIImgList::prevPage(int positions){
 }
 
 /**
- * 
- * @return 
+ *
+ * @return
  */
 bool UIImgList::nextPage(int positions){
     int verticalElems = positions > 0 ? positions : getImgGestor()->calcMaxY(prevImgHeight, this->getH());
     unsigned int posIni = getPosIniLista();
-    
+
     if (verticalElems > 0){
         for (int i=0; i < verticalElems; i++){
             nextRow();
@@ -248,7 +248,7 @@ bool UIImgList::nextPage(int positions){
 }
 
 /**
- * 
+ *
  * @param positions
  */
 bool UIImgList::prevElement(int positions){
@@ -257,7 +257,7 @@ bool UIImgList::prevElement(int positions){
     int total = horizontalElems * verticalElems;
     int allowedJump = getPosActualLista() - positions >= 0 ? positions : 1;
     bool ret = false;
-    
+
     if (horizontalElems > 0){
         if (getPosActualLista() - allowedJump >= 0 && getPosIniLista() > getPosActualLista() - allowedJump){
             setPosActualLista(getPosActualLista() - allowedJump);
@@ -265,7 +265,7 @@ bool UIImgList::prevElement(int positions){
             //setPosFinLista(getPosActualLista() - allowedJump);
             setPosIniLista(getPosFinLista() - total + 1 >= 0 ? getPosFinLista() - total + 1 : 0);
             total = total;
-        } else if (getPosActualLista() >= allowedJump){
+        } else if ((int)getPosActualLista() >= allowedJump){
             setPosActualLista(getPosActualLista() - allowedJump);
             ret = true;
         } else {
@@ -273,12 +273,12 @@ bool UIImgList::prevElement(int positions){
             ret = true;
         }
     }
-    
+
     return ret;
 }
 
 /**
- * 
+ *
  * @param positions
  */
 bool UIImgList::nextElement(int positions){
@@ -287,20 +287,20 @@ bool UIImgList::nextElement(int positions){
     int total = horizontalElems * verticalElems;
     int allowedJump = getPosActualLista() + positions < this->getSize() ? positions : 1;
     bool ret = false;
-    
+
     if (horizontalElems == 1 && getPosActualLista() + allowedJump < this->getSize()){
         setPosActualLista(getPosActualLista() + allowedJump);
         setPosIniLista(getPosActualLista());
         setPosFinLista(getPosActualLista());
     } else if (horizontalElems > 0){
-        if (getPosActualLista() + allowedJump > getPosFinLista() && getPosFinLista() < this->getSize() - 1 
+        if (getPosActualLista() + allowedJump > getPosFinLista() && getPosFinLista() < this->getSize() - 1
                 && getPosActualLista() + allowedJump < this->getSize()){
             setPosActualLista(getPosActualLista() + allowedJump);
             //setPosIniLista( ((getPosActualLista() / horizontalElems) - 1) *  horizontalElems);
             //setPosIniLista( ((getPosActualLista() / horizontalElems)) *  horizontalElems);
             setPosIniLista(getPosIniLista() + horizontalElems);
             setPosFinLista(getPosIniLista() + total - 1 < this->getSize() ? getPosIniLista() + total - 1 : this->getSize() - 1);
-            
+
 //            if (getPosFinLista() == this->getSize() - 1){
 //                setPosIniLista(getPosActualLista() - horizontalElems - allowedJump + 1);
 //            }
@@ -314,13 +314,13 @@ bool UIImgList::nextElement(int positions){
 }
 
 /**
- * 
+ *
  */
 bool UIImgList::nextRow(){
     int horizontalElems = getImgGestor()->calcMaxX(prevImgWidth, this->getW());
     int verticalElems = getImgGestor()->calcMaxY(prevImgHeight, this->getH());
     bool ret = false;
-    
+
     if (horizontalElems > 0){
         int posPantalla = getPosActualLista() - getPosIniLista();
         if (this->getPosActualLista() + horizontalElems < this->getSize()){
@@ -332,7 +332,7 @@ bool UIImgList::nextRow(){
                 setPosActualLista(this->getPosActualLista() + horizontalElems);
                 ret = true;
             }
-        } else if (this->getSize() % horizontalElems > 0 && this->getSize() > horizontalElems * verticalElems){
+        } else if (this->getSize() % horizontalElems > 0 && (int)this->getSize() > horizontalElems * verticalElems){
             setPosFinLista(getSize() - 1);
             setPosActualLista(getPosFinLista());
             int ini = getPosFinLista() - verticalElems * horizontalElems + this->getSize() % horizontalElems + 1;
@@ -343,14 +343,14 @@ bool UIImgList::nextRow(){
 }
 
 /**
- * 
- * @return 
+ *
+ * @return
  */
 bool UIImgList::goToEnd(){
     int horizontalElems = getImgGestor()->calcMaxX(prevImgWidth, this->getW());
     int verticalElems = getImgGestor()->calcMaxY(prevImgHeight, this->getH());
     unsigned int posIni = getPosIniLista();
-    
+
     setPosFinLista(getSize() - 1);
     setPosActualLista(getPosFinLista());
     int ini = getPosFinLista() - verticalElems * horizontalElems + this->getSize() % horizontalElems + 1;
@@ -359,39 +359,39 @@ bool UIImgList::goToEnd(){
 }
 
 /**
- * 
- * @return 
+ *
+ * @return
  */
 bool UIImgList::goToIni(){
     int horizontalElems = getImgGestor()->calcMaxX(prevImgWidth, this->getW());
     int verticalElems = getImgGestor()->calcMaxY(prevImgHeight, this->getH());
     unsigned int posIni = getPosIniLista();
-    
+
     setPosActualLista(0);
     setPosIniLista(0);
-    setPosFinLista(this->getSize() > horizontalElems * verticalElems ? horizontalElems * verticalElems - 1 : this->getSize() - 1);
+    setPosFinLista((int)this->getSize() > horizontalElems * verticalElems ? horizontalElems * verticalElems - 1 : this->getSize() - 1);
     return posIni == getPosIniLista();
 }
 
 /**
- * 
- * @return 
+ *
+ * @return
  */
 bool UIImgList::prevRow(){
     int horizontalElems = getImgGestor()->calcMaxX(prevImgWidth, this->getW());
     int verticalElems = getImgGestor()->calcMaxY(prevImgHeight, this->getH());
     bool ret = false;
     if (horizontalElems > 0){
-        if(getPosActualLista() - horizontalElems < getPosIniLista() && getPosActualLista() > horizontalElems){
+        if((int)getPosActualLista() - horizontalElems < (int)getPosIniLista() && (int)getPosActualLista() > horizontalElems){
            setPosActualLista(getPosActualLista() - horizontalElems);
-           setPosIniLista(getPosIniLista() > horizontalElems ? getPosIniLista() - horizontalElems : 0);
+           setPosIniLista((int)getPosIniLista() > horizontalElems ? getPosIniLista() - horizontalElems : 0);
            setPosFinLista(getPosIniLista() + horizontalElems * verticalElems - 1);
-        } else if (getPosActualLista() <= horizontalElems){
+        } else if ((int)getPosActualLista() <= horizontalElems){
             setPosActualLista(0);
             setPosIniLista(0);
-            setPosFinLista(this->getSize() > horizontalElems * verticalElems ? horizontalElems * verticalElems - 1 : this->getSize() - 1);
+            setPosFinLista((int)this->getSize() > horizontalElems * verticalElems ? horizontalElems * verticalElems - 1 : (int)this->getSize() - 1);
         } else {
-            setPosActualLista(getPosActualLista() - horizontalElems); 
+            setPosActualLista(getPosActualLista() - horizontalElems);
             ret = true;
         }
     }
@@ -399,9 +399,9 @@ bool UIImgList::prevRow(){
 }
 
 /**
- * Evita tener en memoria todas las imagenes de la lista cacheada. 
+ * Evita tener en memoria todas las imagenes de la lista cacheada.
  * Asi se fuerza a recargar las imagenes que se muestran en pantalla en cache
- * 
+ *
  * Esto solo funciona correctamente si se ha podido guardar una miniatura en disco
  * duro. Sino tardaria enormemente luego en volver a recargar las imagenes
  * @param listImages
